get rid of CompositeDocumentViewSelectionFactory
move functionality to DocumentViewsRepository

https://github.com/metasfresh/metasfresh-webui-api/issues/269
