+++ src/main/java/de/metas/ui/web/view/DocumentViewsRepository.java
@@ -2,11 +2,16 @@ package de.metas.ui.web.view;
 
 import java.util.List;
 import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.TimeUnit;
 
+import org.adempiere.exceptions.AdempiereException;
 import org.adempiere.util.Check;
 import org.adempiere.util.lang.impl.TableRecordReference;
+import org.compiere.util.Util.ArrayKey;
+import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Repository;
 
 import com.google.common.base.Preconditions;
@@ -15,8 +20,12 @@ import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.RemovalNotification;
 import com.google.common.collect.ImmutableList;
 
+import de.metas.logging.LogManager;
 import de.metas.ui.web.exceptions.EntityNotFoundException;
 import de.metas.ui.web.view.json.JSONDocumentViewCreateRequest;
+import de.metas.ui.web.view.json.JSONDocumentViewLayout;
+import de.metas.ui.web.view.json.JSONViewDataType;
+import de.metas.ui.web.window.datatypes.json.JSONOptions;
 
 /*
  * #%L
@@ -43,13 +52,64 @@ import de.metas.ui.web.view.json.JSONDocumentViewCreateRequest;
 @Repository
 public class DocumentViewsRepository implements IDocumentViewsRepository
 {
+	private static final Logger logger = LogManager.getLogger(DocumentViewsRepository.class);
+
+	private final ConcurrentHashMap<ArrayKey, IDocumentViewSelectionFactory> factories = new ConcurrentHashMap<>();
 	@Autowired
-	private CompositeDocumentViewSelectionFactory documentViewSelectionFactory;
+	private SqlDocumentViewSelectionFactory defaultFactory;
 
 	private final Cache<String, IDocumentViewSelection> views = CacheBuilder.newBuilder()
 			.expireAfterAccess(1, TimeUnit.HOURS)
 			.removalListener(notification -> onViewRemoved(notification))
 			.build();
+
+	@Autowired
+	public DocumentViewsRepository(final ApplicationContext context)
+	{
+		//
+		// Discover context factories
+		for (final Object factoryObj : context.getBeansWithAnnotation(DocumentViewFactory.class).values())
+		{
+			final IDocumentViewSelectionFactory factory = (IDocumentViewSelectionFactory)factoryObj;
+			final DocumentViewFactory annotation = factoryObj.getClass().getAnnotation(DocumentViewFactory.class);
+			registerFactory(annotation.windowId(), annotation.viewType(), factory);
+		}
+	}
+
+	public void registerFactory(final int windowId, final JSONViewDataType viewType, final IDocumentViewSelectionFactory factory)
+	{
+		Preconditions.checkNotNull(factory, "factory is null");
+		factories.put(mkFactoryKey(windowId, viewType), factory);
+		logger.info("Registered {} for windowId={}, viewType={}", factory, windowId, viewType);
+	}
+
+	private final IDocumentViewSelectionFactory getFactory(final int adWindowId, final JSONViewDataType viewType)
+	{
+		IDocumentViewSelectionFactory factory = factories.get(mkFactoryKey(adWindowId, viewType));
+		if (factory != null)
+		{
+			return factory;
+		}
+
+		factory = factories.get(mkFactoryKey(adWindowId, null));
+		if (factory != null)
+		{
+			return factory;
+		}
+
+		return defaultFactory;
+	}
+
+	private static final ArrayKey mkFactoryKey(final int adWindowId, final JSONViewDataType viewType)
+	{
+		return ArrayKey.of(adWindowId <= 0 ? 0 : adWindowId, viewType);
+	}
+
+	@Override
+	public JSONDocumentViewLayout getViewLayout(final int adWindowId, final JSONViewDataType viewDataType, final JSONOptions jsonOpts)
+	{
+		return getFactory(adWindowId, viewDataType).getViewLayout(adWindowId, viewDataType, jsonOpts);
+	}
 
 	@Override
 	public List<IDocumentViewSelection> getViews()
@@ -60,8 +120,19 @@ public class DocumentViewsRepository implements IDocumentViewsRepository
 	@Override
 	public IDocumentViewSelection createView(final JSONDocumentViewCreateRequest jsonRequest)
 	{
-		final IDocumentViewSelection view = documentViewSelectionFactory.createView(jsonRequest);
+		final int adWindowId = jsonRequest.getAD_Window_ID();
+		final JSONViewDataType viewType = jsonRequest.getViewType();
+		final IDocumentViewSelectionFactory factory = getFactory(adWindowId, viewType);
+		final IDocumentViewSelection view = factory.createView(jsonRequest);
+		if (view == null)
+		{
+			throw new AdempiereException("Failed creating view")
+					.setParameter("request", jsonRequest)
+					.setParameter("factory", factory.toString());
+		}
+
 		views.put(view.getViewId(), view);
+
 		return view;
 	}
 
@@ -69,7 +140,7 @@ public class DocumentViewsRepository implements IDocumentViewsRepository
 	public IDocumentViewSelection getView(final String viewId)
 	{
 		Preconditions.checkNotNull(viewId, "viewId cannot be null");
-		
+
 		final IDocumentViewSelection view = views.getIfPresent(viewId);
 		if (view == null)
 		{
